# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type TaskCategory = typeof TaskCategory[keyof typeof TaskCategory];


export const TaskCategory = {
  DSA: 'DSA',
  Academics: 'Academics',
  Projects: 'Projects',
  Placement: 'Placement',
  Personal: 'Personal',
} as const;

export type TaskPriority = typeof TaskPriority[keyof typeof TaskPriority];


export const TaskPriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export type TaskStatus = typeof TaskStatus[keyof typeof TaskStatus];


export const TaskStatus = {
  all: 'all',
  active: 'active',
  completed: 'completed',
} as const;

export interface Task {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  category: TaskCategory;
  priority: TaskPriority;
  /** @nullable */
  dueDate?: string | null;
  completed: boolean;
  createdAt: string;
  updatedAt: string;
  /** @nullable */
  completedAt?: string | null;
}

export interface TaskInput {
  /**
     * @minLength 1
     * @maxLength 160
     */
  title: string;
  /** @maxLength 1000 */
  description?: string;
  category: TaskCategory;
  priority: TaskPriority;
  dueDate?: string;
}

export interface TaskUpdate {
  /**
     * @minLength 1
     * @maxLength 160
     */
  title?: string;
  /** @maxLength 1000 */
  description?: string;
  category?: TaskCategory;
  priority?: TaskPriority;
  /** @nullable */
  dueDate?: string | null;
  completed?: boolean;
}

export interface CategorySummary {
  category: TaskCategory;
  total: number;
  completed: number;
  active: number;
}

export type UpcomingTask = Task;

export interface DashboardSummary {
  total: number;
  completed: number;
  active: number;
  completionRate: number;
  highPriorityActive: number;
  categorySummaries: CategorySummary[];
  upcoming: Task[];
}

export interface ErrorResponse {
  error: string;
}

export type ListTasksParams = {
category?: TaskCategory;
status?: TaskStatus;
search?: string;
sort?: ListTasksSort;
};

export type ListTasksSort = typeof ListTasksSort[keyof typeof ListTasksSort];


export const ListTasksSort = {
  priority: 'priority',
  dueDate: 'dueDate',
  newest: 'newest',
} as const;