# lib/api-zod/src/generated/types/categorySummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TaskCategory } from './taskCategory';

export interface CategorySummary {
  category: TaskCategory;
  total: number;
  completed: number;
  active: number;
}