# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CategorySummary } from './categorySummary';
import type { Task } from './task';

export interface DashboardSummary {
  total: number;
  completed: number;
  active: number;
  completionRate: number;
  highPriorityActive: number;
  categorySummaries: CategorySummary[];
  upcoming: Task[];
}