# lib/api-zod/src/generated/types/task.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TaskCategory } from './taskCategory';
import type { TaskPriority } from './taskPriority';

export interface Task {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  category: TaskCategory;
  priority: TaskPriority;
  /** @nullable */
  dueDate?: Date | null;
  completed: boolean;
  createdAt: Date;
  updatedAt: Date;
  /** @nullable */
  completedAt?: Date | null;
}