# lib/api-zod/src/generated/types/taskCategory.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type TaskCategory = typeof TaskCategory[keyof typeof TaskCategory];


export const TaskCategory = {
  DSA: 'DSA',
  Academics: 'Academics',
  Projects: 'Projects',
  Placement: 'Placement',
  Personal: 'Personal',
} as const;