# lib/api-zod/src/generated/types/taskStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type TaskStatus = typeof TaskStatus[keyof typeof TaskStatus];


export const TaskStatus = {
  all: 'all',
  active: 'active',
  completed: 'completed',
} as const;