# lib/api-zod/src/generated/types/taskUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TaskCategory } from './taskCategory';
import type { TaskPriority } from './taskPriority';

export interface TaskUpdate {
  /**
     * @minLength 1
     * @maxLength 160
     */
  title?: string;
  /** @maxLength 1000 */
  description?: string;
  category?: TaskCategory;
  priority?: TaskPriority;
  /** @nullable */
  dueDate?: Date | null;
  completed?: boolean;
}